% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_escape}
\alias{add_escape}
\title{Escape sep and target in regex}
\usage{
add_escape(target, sep, split = TRUE)
}
\arguments{
\item{target}{Character string containing a target word}

\item{sep}{Separator in target and lexicon}

\item{split}{Whether to split the target based on sep
and escape every element in the character list}
}
\value{
escaped sep and target
}
\description{
Escape sep and target in regex
}
\keyword{internal}
