% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_input}
\alias{check_input}
\title{Check input and raise warnings as needed}
\usage{
check_input(target, lexicon, sep)
}
\arguments{
\item{target}{Character string containing a target word}

\item{lexicon}{Character vector containing the lexical database}

\item{sep}{Separator in target and lexicon}
}
\value{
the indexes of the competitors in the lexical database
}
\description{
Check input and raise warnings as needed
}
\keyword{internal}
