% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitors.R
\name{get_cohorts}
\alias{get_cohorts}
\title{Get cohort competitors}
\usage{
get_cohorts(
  target,
  lexicon,
  sep = " ",
  form = FALSE,
  count = FALSE,
  overlap = 2
)
}
\arguments{
\item{target}{Character string containing a target word}

\item{lexicon}{Character vector containing the lexical database}

\item{sep}{Separator in target and lexicon}

\item{form}{Whether to return words in lexicon}

\item{count}{Whether to return count of words}

\item{overlap}{(\emph{get_cohorts} only) Integer specifying the number of
onset phonemes to overlap for matching with the target word}
}
\value{
the indexes of the competitors in the lexical database
}
\description{
Cohorts overlap in onset phoneme(s).
}
\examples{
get_cohorts("AA R K", c("AA R K", "AA R T", "B AA B"))
}
