% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitors.R
\name{get_embeds_in_target}
\alias{get_embeds_in_target}
\title{Get embedding competitors}
\usage{
get_embeds_in_target(target, lexicon, sep = " ", form = FALSE, count = FALSE)
}
\arguments{
\item{target}{Character string containing a target word}

\item{lexicon}{Character vector containing the lexical database}

\item{sep}{Separator in target and lexicon}

\item{form}{Whether to return words in lexicon}

\item{count}{Whether to return count of words}
}
\value{
the indexes of the competitors in the lexical database
}
\description{
Embedding competitors are items embedded in target
}
\examples{
get_embeds_in_target("AA R K", c("AA R K", "AA R", "B AA B"))
}
