% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{get_embeds_in_targetP}
\alias{get_embeds_in_targetP}
\title{Get embeds-in-target PRIME}
\usage{
get_embeds_in_targetP(
  target,
  lexicon,
  neighbors = "das",
  sep = " ",
  form = FALSE,
  count = FALSE
)
}
\arguments{
\item{target}{Character string containing a target word}

\item{lexicon}{Character vector containing the lexical database}

\item{neighbors}{(\emph{get_neighbors} only) Character vector specifying the
type of neighbor to return. Return the delete, add, substitute neighbors of
the target when 'd', 'a', and/or 's' is in neighbors respectively}

\item{sep}{Separator in target and lexicon}

\item{form}{Whether to return words in lexicon}

\item{count}{Whether to return count of words}
}
\value{
the indexes of the competitors in the lexical database
}
\description{
Items embedded in the target which are not cohorts or neighbors
}
\examples{
get_embeds_in_targetP("B AA R K IY", c("AA R K", "AA R", "AA R K IY", "B AA R"))
}
