% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{get_fw}
\alias{get_fw}
\title{Get the log Frequency Weight (FW) of a competitor set}
\usage{
get_fw(competitors_freq, pad = 0)
}
\arguments{
\item{competitors_freq}{Numeric vector containing the frequencies of
competitors (including itself)}

\item{pad}{Value to add to frequencies before taking log; if your
minimum frequency is 0, consider adding a value between 1 and 2;
if your minimum frequency is between 0 and 1, consider adding 1}
}
\value{
FW
}
\description{
Get the log Frequency Weight (FW) of a competitor set
}
\examples{
get_fw(c(10, 50), pad = 1)
}
