% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitors.R
\name{get_generic}
\alias{get_generic}
\title{Get competitors in spoken word recognition}
\usage{
get_generic(target, lexicon, neighbors, sep, form, count, overlap)
}
\arguments{
\item{target}{Character string containing a target word}

\item{lexicon}{Character vector containing the lexical database}

\item{neighbors}{(\emph{get_neighbors} only) Character vector specifying the
type of neighbor to return. Return the delete, add, substitute neighbors of
the target when 'd', 'a', and/or 's' is in neighbors respectively}

\item{sep}{Separator in target and lexicon}

\item{form}{Whether to return words in lexicon}

\item{count}{Whether to return count of words}

\item{overlap}{(\emph{get_cohorts} only) Integer specifying the number of
onset phonemes to overlap for matching with the target word}

\item{mismatch}{(\emph{get_rhymes} only) Integer specifying the number of
onset phonemes to mismatch for matching with the target word}
}
\value{
the indexes of the competitors in the lexical database
}
\description{
Get competitors in spoken word recognition
}
\keyword{internal}
