% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitors.R
\name{get_homoforms}
\alias{get_homoforms}
\title{Get homophones}
\usage{
get_homoforms(target, lexicon, sep = " ", form = FALSE, count = FALSE)
}
\arguments{
\item{target}{Character string containing a target word}

\item{lexicon}{Character vector containing the lexical database}

\item{sep}{Separator in target and lexicon}

\item{form}{Whether to return words in lexicon}

\item{count}{Whether to return count of words}
}
\value{
the indexes of the competitors in the lexical database
}
\description{
Homophones are items which sound similar to the target
}
\examples{
get_homoforms("AA R K", c("AA R K", "AA R", "B AA B"))
}
