% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{get_nohorts}
\alias{get_nohorts}
\title{Get nohorts}
\usage{
get_nohorts(
  target,
  lexicon,
  neighbors = "das",
  sep = " ",
  form = FALSE,
  count = FALSE
)
}
\arguments{
\item{target}{Character string containing a target word}

\item{lexicon}{Character vector containing the lexical database}

\item{neighbors}{(\emph{get_neighbors} only) Character vector specifying the
type of neighbor to return. Return the delete, add, substitute neighbors of
the target when 'd', 'a', and/or 's' is in neighbors respectively}

\item{sep}{Separator in target and lexicon}

\item{form}{Whether to return words in lexicon}

\item{count}{Whether to return count of words}
}
\value{
the indexes of the competitors in the lexical database
}
\description{
Items which are both cohorts and neighbors
}
\examples{
get_nohorts("AA R K", c("AA R K", "AA R", "B AA B"), neighbors = "das")
}
