% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_return}
\alias{get_return}
\title{Return lexical items the appropriate type}
\usage{
get_return(idx, lexicon, form, count)
}
\arguments{
\item{lexicon}{Character vector containing the lexical database}

\item{form}{Whether to return words in lexicon}

\item{count}{Whether to return count of words}
}
\value{
the appropriate type
}
\description{
Return lexical items the appropriate type
}
\keyword{internal}
