% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitors.R
\name{get_target_embeds_in}
\alias{get_target_embeds_in}
\title{Get embedded competitors}
\usage{
get_target_embeds_in(target, lexicon, sep = " ", form = FALSE, count = FALSE)
}
\arguments{
\item{target}{Character string containing a target word}

\item{lexicon}{Character vector containing the lexical database}

\item{sep}{Separator in target and lexicon}

\item{form}{Whether to return words in lexicon}

\item{count}{Whether to return count of words}
}
\value{
the indexes of the competitors in the lexical database
}
\description{
Embedded competitors are items which the target embedded in.
}
\examples{
get_target_embeds_in("AA R K", c("AA R K", "B AA R K", "B AA B"))
}
