% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slex}
\alias{slex}
\title{slex ARPAbet}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 212 rows and 3 columns.
}
\source{
\url{https://era.library.ualberta.ca/items/61319cc6-436a-428c-b960-545bdc9bd5d3}
}
\usage{
slex
}
\description{
TRACE slex lexicon translated by Nenadić and Tucker into ARPAbet
pronunciation
}
\details{
TRACE slex lexicon with Frequencies: McClelland, J. L., & Elman, J. L.
(1986). The TRACE model of speech perception. Cognitive psychology, 18(1),
1-86.

APRAbet transcription: Nenadić, F., & Tucker, B. V. (2020). Computational
modelling of an auditory lexical decision experiment using jTRACE and TISK.
Language, Cognition and Neuroscience, 1-29.

@format A table with 212 rows and 2 variables:
\describe{
\item{Item}{TRACE slex transcription}
\item{Pronunciation}{ARPAbet transcription}
...
}
}
\keyword{datasets}
